+++ packages/0x.js/src/contract_wrappers/contract_wrapper.ts
@@ -32,7 +32,7 @@ const CONTRACT_NAME_TO_NOT_FOUND_ERROR: {
 
 export class ContractWrapper {
     protected _web3Wrapper: Web3Wrapper;
-    private _networkId: number;
+    protected _networkId: number;
     private _abiDecoder?: AbiDecoder;
     private _blockAndLogStreamerIfExists?: BlockAndLogStreamer;
     private _blockAndLogStreamIntervalIfExists?: NodeJS.Timer;